+++ RNG_Test/Tests/Integer_Testing/Signed_Short/Signed_Short_Values_Outside_Range.cs
@@ -14,20 +14,20 @@ namespace RNG_Test
 {
 	partial class Integer_Testing
 	{
-		private void Signed_Integer_Values_Outside_Range()
+		private void Signed_Short_Values_Outside_Range()
 		/*
 		===============================================================================================
 		PURPOSE:
-		Test the Random Number Generator for and Signed Integer when the parameters that are provided
-		are outside the range of an Signed Integer.
+		Test the Random Number Generator for and Signed Short when the parameters that are provided
+		are outside the range of a Signed Short.
 		-----------------------------------------------------------------------------------------------
 		OUTPUT:
 		An error message if the values are outside the range; otherwise a randomly generated Signed
-		Integer value.
+		Short value.
 		-----------------------------------------------------------------------------------------------
 		NOTES:
 		- When a value is passed to the RNG, the appropriate overloaded function will be called.
-		  Because of this, a value outside of a Signed Integer becomes a larger unit depending on the
+		  Because of this, a value outside of a Signed Short becomes a larger unit depending on the
 		  value.  To ensure the RESULTS are not outside the range, the calling process needs to
 		  validate the values PRIOR to calling.
 		===============================================================================================
@@ -37,8 +37,8 @@ namespace RNG_Test
 			// Variables - Standard
 			//=============
 			RNG MyRandomNumber = new RNG();
-			int MinValue = int.MinValue;
-			int MaxValue = int.MaxValue;
+			short MinValue = short.MinValue;
+			short MaxValue = short.MaxValue;
 			bool IsValid = true;  // Assume the value is valid
 
 			//=============
@@ -48,29 +48,22 @@ namespace RNG_Test
 			long Bad_MaxValue = int.MaxValue;
 
 			//=============
-			// Setup Environment
-			//=============
-			// Performing this calculation at declaration causes an overflow error.
-			Bad_MinValue -= 1;
-			Bad_MaxValue += 1;
-
-			//=============
 			// Body
 			//=============
 			Con.WriteLine();
 			Con.WriteLine(@"Pass values outside the range");
 
 			// Validate the minimum value
-			if (Bad_MinValue < int.MinValue)
+			if (Bad_MinValue < short.MinValue)
 			{
-				Con.WriteLine(@"Intended value ({0}) cannot be converted to SIGNED Integer.", Bad_MinValue);
+				Con.WriteLine(@"Intended value ({0}) cannot be converted to SIGNED Short.", Bad_MinValue);
 				IsValid = false;
 			}
 
 			// Validate the maximum value
-			if (Bad_MaxValue > int.MaxValue)
+			if (Bad_MaxValue > short.MaxValue)
 			{
-				Con.WriteLine(@"Intended value ({0}) cannot be converted to SIGNED Integer.", Bad_MaxValue);
+				Con.WriteLine(@"Intended value ({0}) cannot be converted to SIGNED Short.", Bad_MaxValue);
 				IsValid = false;
 			}
 
@@ -83,6 +76,6 @@ namespace RNG_Test
 				Con.Write("{0}", MyRandomNumber.Generate(MinValue, Bad_MaxValue) + "\t");
 				Con.WriteLine("{0}", MyRandomNumber.Generate(Bad_MinValue, Bad_MaxValue));
 			}
-		} // private void Signed_Integer_Values_Outside_Range
+		} // private void Signed_Short_Values_Outside_Range
 	} // partial class Integer_Testing
 } // namespace RNG_Test