+++ RNG_Test/Tests/Integer_Testing/Unsgned_Short/Unsigned_Short_Same_Value.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Orias_RNG;
+
+//=============
+// Aliases
+//=============
+using Con = System.Console;
+
+namespace RNG_Test
+{
+	partial class Integer_Testing
+	{
+		private void Unsigned_Short_Same_Value()
+		/*
+		===============================================================================================
+		PURPOSE:
+		Test the Random Number Generator for an Unsigned Short when the parameters that are provided
+		are the same value.
+		-----------------------------------------------------------------------------------------------
+		OUTPUT:
+		Randomly generated Unsigned Short values.
+		===============================================================================================
+		*/
+		{
+			//=============
+			// Variables - Standard
+			//=============
+			RNG MyRandomNumber = new RNG();
+			ushort MinValue = ushort.MinValue;
+			ushort MaxValue = ushort.MaxValue;
+
+			//=============
+			// Body
+			//=============
+			Con.WriteLine();
+			Con.WriteLine(@"Pass the same value");
+			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MinValue) + "\t");
+			Con.WriteLine("{0}", MyRandomNumber.Generate(MaxValue, MaxValue));
+		} // private void Unsigned_Short_Same_Value
+	} // partial class Integer_Testing
+} // namespace RNG_Test